split function reverse in main into collecting code blocks and printing
